Reduced the number of API accesses and object creation, improving application response time when collecting songs. This is because songs and albums are discarded at the time the name is obtained.
After the process of collecting songs to the database, the songs of an artist are deleted if their popularity is less than or equal to the first quantile of the popularity of all the songs of that artist.

Now, an error message appears if you try to collect songs from an artist that is already in the database; and a confirmation message appears after the collection process ends, also showing the popularity in which the less popular songs were removed.